+++ cli/cmd/testdata/install_ha_output.golden
@@ -247,7 +247,6 @@ rules:
       - grpcroutes/status
     verbs:
       - patch
-      - create
   - apiGroups:
       - workload.linkerd.io
     resources:
@@ -1411,7 +1410,7 @@ spec:
   template:
     metadata:
       annotations:
-        checksum/config: 46e971b5b0cf8f212d9c50d0a90ca19ec6875216f3420f259e5e2fa09744cd77
+        checksum/config: 92d3d7185abd241437f7889dabe15a840b3b91f8910f9491ab10de3a1074ed6d
         linkerd.io/created-by: linkerd/cli dev-undefined
         linkerd.io/proxy-version: install-proxy-version
         cluster-autoscaler.kubernetes.io/safe-to-evict: "true"
