+++ cli/cmd/testdata/install_helm_control_plane_output.golden
@@ -238,7 +238,6 @@ rules:
       - grpcroutes/status
     verbs:
       - patch
-      - create
   - apiGroups:
       - workload.linkerd.io
     resources:
@@ -1288,7 +1287,7 @@ spec:
   template:
     metadata:
       annotations:
-        checksum/config: c5e0cd8a3d2c01086bbf24974ff50ad866387850e049f8283e2c562d0f461a7b
+        checksum/config: aae2f97af199e93464b02da923a0732b603488c822241c14b2d86ba2bf6da03d
         linkerd.io/created-by: linkerd/helm linkerd-version
         linkerd.io/proxy-version: test-proxy-version
         cluster-autoscaler.kubernetes.io/safe-to-evict: "true"
