+++ cli/cmd/testdata/install_helm_output_ha_namespace_selector.golden
@@ -238,7 +238,6 @@ rules:
       - grpcroutes/status
     verbs:
       - patch
-      - create
   - apiGroups:
       - workload.linkerd.io
     resources:
@@ -1376,7 +1375,7 @@ spec:
   template:
     metadata:
       annotations:
-        checksum/config: 68a532538d399548559f765b2291422ff5e91ee122c6d8c43fa09b19495b38f1
+        checksum/config: 92e9069580ccc3e2e6a5a23fe9d4586ae8f0f3b48b906119e6f884ff66ea7395
         linkerd.io/created-by: linkerd/helm linkerd-version
         linkerd.io/proxy-version: test-proxy-version
         cluster-autoscaler.kubernetes.io/safe-to-evict: "true"
