+++ cli/cmd/testdata/install_proxy_ignores.golden
@@ -247,7 +247,6 @@ rules:
       - grpcroutes/status
     verbs:
       - patch
-      - create
   - apiGroups:
       - workload.linkerd.io
     resources:
@@ -1313,7 +1312,7 @@ spec:
   template:
     metadata:
       annotations:
-        checksum/config: c5587f7ed43f6c2e7aa3d797837dd88962c632b695cb2380436487abbc6fc13c
+        checksum/config: 7ed64a2dea184e1e880a33bb5577b3f6c2a3fdf72492321683bbdaef30a1f365
         linkerd.io/created-by: linkerd/cli dev-undefined
         linkerd.io/proxy-version: install-proxy-version
         cluster-autoscaler.kubernetes.io/safe-to-evict: "true"
